fix: Improve app lifecycle and task switcher interaction
This separates the menu bar and browser window into two separate apps with the menu bar app embedded in the main Reconnect browser app.

This change likely introduces a few short-term lifecycle issues, but will hopefully vastly improve how the main app interacts with the task switcher as it no longer needs to dynamically hide and show the Dock icon.
