+++ Reconnect Menu/Assets.xcassets/StatusConnected.imageset/Contents.json
