+++ Reconnect Menu/Assets.xcassets/StatusConnected.imageset/StatusConnectedDark.png
