+++ Reconnect Menu/Assets.xcassets/StatusDisconnected.imageset/StatusDisconnected.png
