+++ Reconnect Menu/Assets.xcassets/StatusDisconnected.imageset/StatusDisconnectedDark.png
