+++ ReconnectCore/Package.swift
@@ -5,6 +5,9 @@ import PackageDescription
 
 let package = Package(
     name: "ReconnectCore",
+    platforms: [
+        .macOS(.v12),
+    ],
     products: [
         // Products define the executables and libraries a package produces, making them visible to other packages.
         .library(
